+++ src/testGradeData/TestUpdateWeightedGrades.java
@@ -22,11 +22,11 @@ public class TestUpdateWeightedGrades {
 		grade_data.setStudentList(list);
 		grade_data.updateWeightedGrades();
 		
-		int expect1 = Math.round(69*0.2f + 77*0.2f + 55*0.2f + 0*0.2f + 18*0.2f);
+		int expect1 = (int) Math.round(69*0.2 + 77*0.2 + 55*0.2 + 0*0.2 + 18*0.2);
 		assertEquals(expect1, list.get(0).getWeightedGrade());
-		int expect2 = Math.round(56*0.2f + 89*0.2f + 85*0.2f + 88*0.2f + 87*0.2f);
+		int expect2 = (int) Math.round(56*0.2 + 89*0.2 + 85*0.2 + 88*0.2 + 87*0.2);
 		assertEquals(expect2, list.get(1).getWeightedGrade());
-		int expect3 = Math.round(52*0.2f + 0*0.2f + 13*0.2f + 14*0.2f + 99*0.2f);
+		int expect3 = (int) Math.round(52*0.2 + 0*0.2 + 13*0.2 + 14*0.2 + 99*0.2);
 		assertEquals(expect3, list.get(2).getWeightedGrade());
 	}
 
@@ -39,14 +39,14 @@ public class TestUpdateWeightedGrades {
 		Student s3 = new Student("105062273", "¦ó¤j«L", new int[]{9, 8, 7, 6, 5});
 		ArrayList<Student> list = new ArrayList<Student>(Arrays.asList(s1, s2, s3));
 		grade_data.setStudentList(list);
-		Student.setWeights(new float[]{0.1f, 0.05f, 0.375f, 0.1f, 0.375f});
+		Student.setWeights(new double[]{0.1, 0.05, 0.375, 0.1, 0.375});
 		grade_data.updateWeightedGrades();
 		
-		int expect1 = Math.round(23*0.1f + 15*0.05f + 100*0.375f + 100*0.1f + 2*0.375f);
+		int expect1 = (int) Math.round(23*0.1 + 15*0.05 + 100*0.375 + 100*0.1 + 2*0.375);
 		assertEquals(expect1, list.get(0).getWeightedGrade());
-		int expect2 = Math.round(66*0.1f + 89*0.05f + 77*0.375f + 12*0.1f + 16*0.375f);
+		int expect2 = (int) Math.round(66*0.1 + 89*0.05 + 77*0.375 + 12*0.1 + 16*0.375);
 		assertEquals(expect2, list.get(1).getWeightedGrade());
-		int expect3 = Math.round(9*0.1f + 8*0.05f + 7*0.375f + 6*0.1f + 5*0.375f);
+		int expect3 = (int) Math.round(9*0.1 + 8*0.05 + 7*0.375 + 6*0.1 + 5*0.375);
 		assertEquals(expect3, list.get(2).getWeightedGrade());
 	}
